+++ CHANGELOG.md
@@ -1,1 +1,29 @@
 # Changelog
+
+## 0.3.0 (19.03.2024)
+- [Add Component - Email fetcher](https://github.com/kommitters/bns/issues/40)
+- [Use case - Email based notifications](https://github.com/kommitters/bns/issues/45)
+- [Dynamic WIP limit value](https://github.com/kommitters/bns/issues/48)
+
+## 0.2.0 (29.02.2024)
+- [Add a Postgres fetcher component](https://github.com/kommitters/bns/issues/28)
+- [Use case - PTO's Postgres-Slack implementation](https://github.com/kommitters/bns/issues/30)
+- [Slack dispatcher](https://github.com/kommitters/bns/issues/32)
+- [Refactor fetcher components](https://github.com/kommitters/bns/issues/34)
+- [Add Use Case - Boards WI alerts](https://github.com/kommitters/bns/issues/36)
+- [Add Use Case - Next Week Birthday](https://github.com/kommitters/bns/issues/38)
+- [Add Use Case - Next Week PTO's](https://github.com/kommitters/bns/issues/39)
+
+## 0.1.1 (07.02.2024)
+- [Add custom templates option](https://github.com/kommitters/bns/issues/25)
+- [PTO's formatting use cases](https://github.com/kommitters/bns/issues/24)
+
+## 0.1.0 (06.02.2024)
+
+- [Build initial codebase](https://github.com/kommitters/bns/issues/6)
+- [Birthdays use case implementation](https://github.com/kommitters/bns/issues/7)
+- [PTO’s use case implementation](https://github.com/kommitters/bns/issues/8)
+- [Error handler implementation](https://github.com/kommitters/bns/issues/9)
+- [Types implementation](https://github.com/kommitters/bns/issues/11)
+- [Documentation completion](https://github.com/kommitters/bns/issues/12)
+- [Opensource configurations](https://github.com/kommitters/bns/issues/10)